intellihide: Fix unexpected unredirection disable with overview
When overview shows, this._animateIn is called in
_onOverviewShowing(). And when overview hides, _updateDashVisibility()
might call this._animateIn() again, which makes the unredirection always
disabled and brings performance issue in some cases.

This commit replaces the counter for unredirection by a boolean value,
to make logic more clear, and not disabling/enabling unredirection for
multiple times.